Check if entry exists before trying to delete

+++ src/main/java/com/personal/learning/journal/controller/journalController.java
@@ -56,6 +56,9 @@ public class journalController {
 
     @GetMapping(path = "deleteEntry")
     public String deleteEntry(@RequestParam("entryId") Long entryId){
+        if(!service.existsById(entryId)){
+            return "entryNotExist";
+        }
         service.deleteById(entryId);
         return "entryDeleted";
     }
